# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Controle de Bebidas - SENAI Jandira
  description: |-
   API destinada para controle de bebidas
  termsOfService: http://swagger.io/terms/
  contact:
    email: kayquebrennoferreira@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://
paths:
...
  /v1/delicia-gelada/admin/sabor:
    post:
      tags:
        - Sabor
      summary: Adiciona um novo Sabor
      description: Adiciona um novo Sabor
      requestBody:
        description: JSON contendo os dados do novo SABOR a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sabor'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Sabor
      summary: Lista todos os Sabores
      description: Listagem de todos os Sabores registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaborGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
components:
  schemas:
...
    Sabor:
      required:
        - nome_sabor
      type: object
      properties:
        nome_sabor:
          type: string
          example: Cola
          description: O número máximo de caracteres é de 60
          minLength: 1
          maxLength: 60
...
    SaborGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome_sabor:
          type: string
          example: Cola
          description: O número máximo de caracteres é de 60
...